requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiofiles==24.1.0
    # via sanic
html5tagger==1.3.0
    # via
    #   sanic
    #   tracerite
httptools==0.6.1
    # via sanic
multidict==6.1.0
    # via sanic
python-dotenv==1.0.1
    # via -r requirements.in
sanic==24.6.0
    # via -r requirements.in
sanic-routing==23.12.0
    # via sanic
tracerite==1.1.1
    # via sanic
typing-extensions==4.12.2
    # via sanic
ujson==5.10.0
    # via sanic
uvloop==0.20.0
    # via sanic
websockets==13.1
    # via sanic

# The following packages are considered to be unsafe in a requirements file:
# setuptools
